refactor(Core/Unit): Split spell things into SpellDefines.h
